move projecttracks, verify recaptcha
